pyproject.toml
[tool.poetry]
name = "stock-fund-analysis"
version = "0.1.0"
description = "个股资金流分析应用"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.28.0"
akshare = "^1.17.44"
pandas = "^2.1.0"
numpy = "^2.0.0"
loguru = "^0.7.2"
python-dotenv = "^1.0.0"
openai = "^1.3.0"
requests = "^2.31.0"
plotly = "^5.17.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.9.0"
flake8 = "^6.1.0"
mypy = "^1.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true